회원가입 과정별로 컴포넌트 세분화 (ProgressSteps, Agreements, Inform, Congratulation - 예정) - 라우팅 주소, 폰트 추가 수정
