+++ festive-app/src/AppRoutes.jsx
@@ -1,7 +1,7 @@
 import {Route, Routes} from "react-router-dom";
 import MainPage from "./scomponents/mainPage/mainPage.jsx";
 import Signin from "./jcomponents/Signin/Signin.jsx";
-import SignUp1 from "./jcomponents/Signup/Signup1.jsx";
+import Signup from "./jcomponents/Signup/Signup.jsx";
 import WaglePage from "./ycomponents/WaglePage.jsx";
 import WritePage from "./ycomponents/WritePage.jsx";
 import Find from "./jcomponents/FindId/Find.jsx";
@@ -14,7 +14,7 @@ const AppRoutes = () => {
         <Route path="/" element={<MainPage />} />
         <Route path="/signin" element={<Signin />} />
         <Route path="/find" element={<Find />} />
-        <Route path="/signup" element={<SignUp1 />} />
+        <Route path="/signup/agreement" element={<Signup />} />
         <Route path="/wagle" element={<WaglePage />} />
         <Route path="/wagle/write" element={<WritePage />} />
         <Route path="/wagle/:id" element={<WagleDetail />} />
