+++ festive-app/src/jcomponents/Signin/Signin.css
@@ -50,7 +50,7 @@
 .login-title {
   font-size: 1.5rem; /* text-2xl */
   line-height: 2rem;
-  font-weight: 800; /* font-medium */
+  font-weight: 900; /* font-medium */
   color: #60a5fa; /* text-blue-400 */
   margin: 0;
 }
@@ -113,11 +113,16 @@
 .find-link {
   color: inherit;
   text-decoration: none;
+  font-weight: 400;
   transition: color 0.15s ease-in-out; /* transition-colors */
 }
 
 .find-link:hover {
   color: #60a5fa; /* hover:text-blue-500 */
+}
+
+.separator {
+  font-weight: 400;
 }
 
 /* 버튼 그룹 */
@@ -166,6 +171,7 @@
 
 .sns-title span {
   font-size: 0.875rem; /* text-sm */
+  font-weight: 400;
   line-height: 1.25rem;
   color: #6b7280; /* text-gray-500 */
 }
