+++ festive-app/src/jcomponents/Signup/Signup.css
@@ -1,4 +1,4 @@
-/* SignUp2 Container */
+/* Signup Container */
 .signup-container {
   min-height: 100vh;
   background-color: #f9fafb;
@@ -57,7 +57,7 @@
   width: 4rem;
   height: 2px;
   background-color: #e5e7eb;
-  margin: 0 1rem;
+  margin: 0 1rem 1rem;
 }
 
 /* Main Content */
@@ -72,7 +72,10 @@
   font-size: 1.25rem;
   font-weight: 600;
   color: #374151;
-  margin-bottom: 0.5rem;
+  margin-bottom: 1.5rem;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
 }
 
 .signup-subtitle {
@@ -80,6 +83,166 @@
   color: #6b7280;
   margin-bottom: 2rem;
   margin-top: 0;
+}
+
+/* Terms Sections */
+.terms-section {
+  border: 1px solid #e5e7eb;
+  border-radius: 0.5rem;
+  margin-bottom: 1rem;
+  background-color: #f8fafc;
+}
+
+.terms-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 1rem;
+}
+
+.checkbox-label {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  cursor: pointer;
+  font-size: 0.875rem;
+  color: #374151;
+  font-weight: 500;
+}
+
+.checkbox-label input[type="checkbox"] {
+  display: none;
+}
+
+.checkmark {
+  width: 1.25rem;
+  height: 1.25rem;
+  border: 2px solid #d1d5db;
+  border-radius: 0.25rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 0.2s ease;
+}
+
+.checkbox-label input[type="checkbox"]:checked + .checkmark {
+  background-color: #60a5fa;
+  border-color: #60a5fa;
+}
+
+.checkbox-label input[type="checkbox"]:checked + .checkmark::after {
+  content: "✓";
+  color: white;
+  font-size: 0.75rem;
+  font-weight: bold;
+}
+
+.toggle-btn {
+  background-color: #f3f4f6;
+  border: 1px solid #d1d5db;
+  border-radius: 0.375rem;
+  padding: 0.5rem 1rem;
+  font-size: 0.875rem;
+  color: #6b7280;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.toggle-btn:hover {
+  background-color: #e5e7eb;
+  color: #374151;
+}
+
+.terms-content {
+  height: 100px;
+  overflow-y: auto;
+  padding: 1rem;
+  border-top: 1px solid #e5e7eb;
+  background-color: white;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: #6b7280;
+  line-height: 1.5;
+  white-space: pre-wrap;
+}
+
+/* Bottom Notice */
+.bottom-notice {
+  margin-top: 2rem;
+  padding: 1rem;
+  background-color: #f8fafc;
+  border-radius: 0.5rem;
+  border: 1px solid #e5e7eb;
+  text-align: center;
+}
+
+.bottom-notice p {
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: #6b7280;
+  margin-bottom: 1rem;
+  margin-top: 0;
+}
+
+.select-all-label {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.75rem;
+  cursor: pointer;
+  font-size: 0.875rem;
+  color: #374151;
+  font-weight: 600;
+}
+
+.select-all-label input[type="checkbox"] {
+  display: none;
+}
+
+.select-all-label .checkmark {
+  width: 1.25rem;
+  height: 1.25rem;
+  border: 2px solid #d1d5db;
+  border-radius: 0.25rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 0.2s ease;
+}
+
+.select-all-label input[type="checkbox"]:checked + .checkmark {
+  background-color: #60a5fa;
+  border-color: #60a5fa;
+}
+
+.select-all-label input[type="checkbox"]:checked + .checkmark::after {
+  content: "✓";
+  color: white;
+  font-size: 0.75rem;
+  font-weight: bold;
+}
+
+/* Next Button */
+.step-btn {
+  width: 100%;
+  background-color: #60a5fa;
+  color: white;
+  border: none;
+  border-radius: 0.5rem;
+  padding: 0.75rem;
+  font-size: 1rem;
+  font-weight: 600;
+  cursor: pointer;
+  margin-top: 2rem;
+  transition: background-color 0.2s ease;
+}
+
+.step-btn:hover {
+  background-color: #3b82f6;
+}
+
+.step-btn:disabled {
+  background-color: #d1d5db;
+  cursor: not-allowed;
 }
 
 /* Form Styles */
@@ -128,10 +291,6 @@
 
 .form-input::placeholder {
   color: #9ca3af;
-}
-
-.full-width {
-  width: 100%;
 }
 
 .input-group {
@@ -229,7 +388,7 @@
   
   .step-line {
     width: 2rem;
-    margin: 0 0.5rem;
+    margin: 0 0.5rem 0.5rem;
   }
   
   .form-row {
