--- festive-app/src/jcomponents/Signup/Signup1.css
@@ -1,254 +1,0 @@
-/* SignUp1 Container */
-.signup-container {
-  min-height: 100vh;
-  background-color: #f9fafb;
-  padding: 2rem 1rem;
-}
-
-.signup-wrapper {
-  max-width: 48rem;
-  margin: 0 auto;
-}
-
-/* Progress Steps */
-.progress-steps {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-bottom: 3rem;
-}
-
-.step {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.step-number {
-  width: 2.5rem;
-  height: 2.5rem;
-  border-radius: 50%;
-  background-color: #e5e7eb;
-  color: #6b7280;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-weight: 600;
-  transition: all 0.2s ease;
-}
-
-.step.active .step-number {
-  background-color: #60a5fa;
-  color: white;
-}
-
-.step-label {
-  font-size: 0.875rem;
-  color: #6b7280;
-  font-weight: 500;
-}
-
-.step.active .step-label {
-  color: #60a5fa;
-}
-
-.step-line {
-  width: 4rem;
-  height: 2px;
-  background-color: #e5e7eb;
-  margin: 0 1rem;
-}
-
-/* Main Content */
-.signup-content {
-  background-color: white;
-  border-radius: 0.5rem;
-  padding: 2rem;
-  box-shadow: 0 1px 3px 0 rgba(0, 0, 0, 0.1);
-}
-
-.signup-title {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: #374151;
-  margin-bottom: 1.5rem;
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-/* Terms Sections */
-.terms-section {
-  border: 1px solid #e5e7eb;
-  border-radius: 0.5rem;
-  margin-bottom: 1rem;
-  background-color: #f8fafc;
-}
-
-.terms-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem;
-}
-
-.checkbox-label {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  cursor: pointer;
-  font-size: 0.875rem;
-  color: #374151;
-  font-weight: 500;
-}
-
-.checkbox-label input[type="checkbox"] {
-  display: none;
-}
-
-.checkmark {
-  width: 1.25rem;
-  height: 1.25rem;
-  border: 2px solid #d1d5db;
-  border-radius: 0.25rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.2s ease;
-}
-
-.checkbox-label input[type="checkbox"]:checked + .checkmark {
-  background-color: #60a5fa;
-  border-color: #60a5fa;
-}
-
-.checkbox-label input[type="checkbox"]:checked + .checkmark::after {
-  content: "✓";
-  color: white;
-  font-size: 0.75rem;
-  font-weight: bold;
-}
-
-.toggle-btn {
-  background-color: #f3f4f6;
-  border: 1px solid #d1d5db;
-  border-radius: 0.375rem;
-  padding: 0.5rem 1rem;
-  font-size: 0.875rem;
-  color: #6b7280;
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-.toggle-btn:hover {
-  background-color: #e5e7eb;
-  color: #374151;
-}
-
-.terms-content {
-  padding: 1rem;
-  border-top: 1px solid #e5e7eb;
-  background-color: white;
-  font-size: 0.875rem;
-  color: #6b7280;
-  line-height: 1.5;
-}
-
-/* Bottom Notice */
-.bottom-notice {
-  margin-top: 2rem;
-  padding: 1rem;
-  background-color: #f8fafc;
-  border-radius: 0.5rem;
-  border: 1px solid #e5e7eb;
-  text-align: center;
-}
-
-.bottom-notice p {
-  font-size: 0.875rem;
-  color: #6b7280;
-  margin-bottom: 1rem;
-  margin-top: 0;
-}
-
-.select-all-label {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.75rem;
-  cursor: pointer;
-  font-size: 0.875rem;
-  color: #374151;
-  font-weight: 600;
-}
-
-.select-all-label input[type="checkbox"] {
-  display: none;
-}
-
-.select-all-label .checkmark {
-  width: 1.25rem;
-  height: 1.25rem;
-  border: 2px solid #d1d5db;
-  border-radius: 0.25rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.2s ease;
-}
-
-.select-all-label input[type="checkbox"]:checked + .checkmark {
-  background-color: #60a5fa;
-  border-color: #60a5fa;
-}
-
-.select-all-label input[type="checkbox"]:checked + .checkmark::after {
-  content: "✓";
-  color: white;
-  font-size: 0.75rem;
-  font-weight: bold;
-}
-
-/* Next Button */
-.next-btn {
-  width: 100%;
-  background-color: #60a5fa;
-  color: white;
-  border: none;
-  border-radius: 0.5rem;
-  padding: 0.75rem;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  margin-top: 2rem;
-  transition: background-color 0.2s ease;
-}
-
-.next-btn:hover {
-  background-color: #3b82f6;
-}
-
-.next-btn:disabled {
-  background-color: #d1d5db;
-  cursor: not-allowed;
-}
-
-/* Responsive */
-@media (max-width: 768px) {
-  .signup-container {
-    padding: 1rem 0.5rem;
-  }
-  
-  .signup-content {
-    padding: 1.5rem;
-  }
-  
-  .progress-steps {
-    margin-bottom: 2rem;
-  }
-  
-  .step-line {
-    width: 2rem;
-    margin: 0 0.5rem;
-  }
-}
